fix: modify edit and remove in calendar component
