+++ packages/client/src/schema-component/antd/calendar/index.md
@@ -14,7 +14,3 @@ group:
 #### Sync Data
 
 <code src="./demos/demo1.tsx" />
-
-#### async Data
-
-<code src="./demos/demo2.tsx" />
